+++ src/main/java/kr/anabada/anabadaserver/domain/change/respository/ProductRepository.java
@@ -3,5 +3,5 @@ package kr.anabada.anabadaserver.domain.change.respository;
 import kr.anabada.anabadaserver.domain.change.entity.MyProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MyProductRepository extends JpaRepository<MyProduct, Long>, MyProductRepositoryCustom {
+public interface ProductRepository extends JpaRepository<MyProduct, Long>, ProductRepositoryCustom {
 }
